python/src/api/uploadcsv/requirements.txt
astroid==2.15.6
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
dill==0.3.7
dnspython==2.4.2
easygui==0.98.3
Flask==2.3.2
Flask-MySQL==1.5.2
Flask-Navigation==0.2.0
Flask-PyMongo==2.3.0
grpcio==1.58.0
grpcio-tools==1.58.0
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jsonify==0.5
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
parso==0.8.3
pika==1.3.2
platformdirs==3.10.0
protobuf==4.21.12
pymongo==4.5.0
pycparser==2.21
pylint==2.17.5
PyMySQL==1.1.0
PyJWT==2.8.0
requests==2.31.0
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
wrapt==1.15.0
